Adding Default Git Commit Template
Related To Issue: N/A

What's Wrong: No default commit template.

How Was it Fixed (if not obvious):
* Add template
* Change git config

If this does not show up for you add this to .git/config
```
[commit]
        template = documentation/commit.txt
```

What side effects does this have (could be none):
* Better commit messages ideally

Which builds did you run to make sure they build?

[ ] arm-none-eabi-gcc Cortex M4
[ ] arm-none-eabi-gcc Cortex M7
[ ] (Apple) Native Clang
[ ] (Apple) Homebrew GCC
[ ] (Apple) Homebrew LLVM

How Do We Know and Can Show It's Fixed:

* This commit template showed up in VSCode.

Which Unittest Series did you Check?

[ ] (Apple) Native Clang
[ ] (Apple) Homebrew GCC
[ ] (Apple) Homebrew LLVM

Did this affect any on-target builds? If so which were tested?

[ ] STM32F407VE board
[ ] STM32H753ZI board