Implement Cache API through SVC calls
Related To Issue: 28

What's Wrong:

* Cache operations are not supported in SVC calls.

How Was it Fixed (if not obvious):

* Adding Cache API to jarnax
* Reformatting SVC marshal to handle calls as more topical areas
* Individual operations on topical areas are passed differently.

What side effects does this have (could be none):

* Other APIS have to be ported as well in the future (Ticks, etc)

Which builds did you run to make sure they build?

[x] arm-none-eabi-gcc Cortex M4
[x] arm-none-eabi-gcc Cortex M7
[x] (Apple) Native Clang
[ ] (Apple) Homebrew GCC
[x] (Apple) Homebrew LLVM

How Do We Know and Can Show It's Fixed:

* We can not yet (on a cruise)

Which Unittest Series did you Check?

[x] (Apple) Native Clang
[ ] (Apple) Homebrew GCC
[ ] (Apple) Homebrew LLVM

Did this affect any on-target builds? If so which were tested?
Affects both, untested as of yet.

[ ] STM32F407VE board
[ ] STM32H753ZI board