LLVM/Clang build should be more similar
What's Wrong:

* Clang and LLVM build dissimilarities are growing and should be kept close.

How Was it Fixed (if not obvious):

* Make sure they build in a similar way and make similar assumptions. They should not be able to assume gcc as a compiler for example.

What side effects does this have (could be none):

* None

Which builds did you run to make sure they build?

[ ] arm-none-eabi-gcc Cortex M4
[ ] arm-none-eabi-gcc Cortex M7
[X] (Apple) Native Clang
[ ] (Apple) Homebrew GCC
[X] (Apple) Homebrew LLVM

How Do We Know and Can Show It's Fixed:

* Builds

Which Unittest Series did you Check?

[X] (Apple) Native Clang
[X] (Apple) Homebrew GCC
[X] (Apple) Homebrew LLVM

Did this affect any on-target builds? If so which were tested?

[ ] STM32F407VE board
[ ] STM32H753ZI board