Adding unit tests for Buffer beyond the existing Heap tests.
What's Wrong:

* Buffer was under-tested

How Was it Fixed (if not obvious):

* Copilot w/ Claude to add unit tests.

What side effects does this have (could be none):

* More unit tests

Which builds did you run to make sure they build?

[ ] arm-none-eabi-gcc Cortex M4
[ ] arm-none-eabi-gcc Cortex M7
[ ] (Apple) Native Clang
[X] (Apple) Homebrew GCC
[ ] (Apple) Homebrew LLVM

How Do We Know and Can Show It's Fixed:

* Unit tests pass

Which Unittest Series did you Check?

[ ] (Apple) Native Clang
[X] (Apple) Homebrew GCC
[ ] (Apple) Homebrew LLVM

Did this affect any on-target builds? If so which were tested?

[ ] STM32F407VE board
[ ] STM32H753ZI board