Adding Tests for Conversions
What's Wrong:

* Conversion tests are missing.

How Was it Fixed (if not obvious):

* Added unit tests w/ Copilot /w Claude

What side effects does this have (could be none):

* None

Which builds did you run to make sure they build?

[ ] arm-none-eabi-gcc Cortex M4
[ ] arm-none-eabi-gcc Cortex M7
[ ] (Apple) Native Clang
[X] (Apple) Homebrew GCC
[ ] (Apple) Homebrew LLVM

How Do We Know and Can Show It's Fixed:

* Unit tests pass

Which Unittest Series did you Check?

[ ] (Apple) Native Clang
[X] (Apple) Homebrew GCC
[ ] (Apple) Homebrew LLVM

Did this affect any on-target builds? If so which were tested?

[ ] STM32F407VE board
[ ] STM32H753ZI board