Adding more functionality to the Variant class.
What's Wrong:

* Variants don't have great constructor or get methods

How Was it Fixed (if not obvious):

* Asked Copilot w/ Claude to add these and add unit tests.

What side effects does this have (could be none):

* Variant usage in the code base should go up.

Which builds did you run to make sure they build?

[ ] arm-none-eabi-gcc Cortex M4
[ ] arm-none-eabi-gcc Cortex M7
[ ] (Apple) Native Clang
[X] (Apple) Homebrew GCC
[ ] (Apple) Homebrew LLVM

How Do We Know and Can Show It's Fixed:

* Unit tests pass

Which Unittest Series did you Check?

[ ] (Apple) Native Clang
[X] (Apple) Homebrew GCC
[ ] (Apple) Homebrew LLVM

Did this affect any on-target builds? If so which were tested?

[ ] STM32F407VE board
[ ] STM32H753ZI board